Forecasting Company Revenue
This report deals with the analysis of monthly data for a small web based company that makes money from advertisement posting on its web pages. The purpose of this study is to analyze historical data and come up with the most appropriate forecast of total earnings for the first quarter of January, 2017. We aim to predict the website’s total earnings for the above mentioned period by analyzing one year of monthly revenue records from January 2016 to December, 2016.